Remove leading track number if present in file name fallback
Prepare RE rules parsing from a string variable in addition to the
above as such option can become an user setting in the future.
This is vastly ugly to use because Lua's *table* isn't available
as a config option value in mpv config files. Because of that this
stays hidden and hardcoded for now.

Don't replace underscores '_' with spaces ' ' for folder names
too.