Remove old generative template variables and related code
The text areas 1, 2 and 3 are already configurable via template
variables, therefore there's no need for a temporary solution
based on generative template content (GEN).

This patch removes support for the following template variables
which are no longer parsed specially:
  TEXTAREA_1_GEN
  TEXTAREA_2_GEN
  TEXTAREA_3_GEN

Text area 4 is still filled in with a pre-defined generative
variable, therefore the last template variable still stays
present:
  TEXTAREA_4_GEN

Users can replace the generative variables *_GEN with the default
configuration values from the example configuration file. The
pre-defined configuration is the same as it was with the
generative variables.