Persist user-toggled OSD states
If user manually hides or activates an OSD (OSD-1 or OSD-2),
persist the state on chapter changes and track change events, that
is on loading a media from the playlist.

Update remote source detection algorithm:

Remote source can be either a HTTP Live Stream (hls) or a media
with URI path in format 'PROT://PATH'.

Add file date tags to remote source template:

File date tags are broken into individual date fields and added to
the remote source template in format YYYY-MM-DD.

If the raw date format is a persistent field or not, and if it may
change in the future remains unanswered as for the current patch
release date, as well as a choice of a pre-defined date format.

The file date tag can removed from the template in user config
file by setting:
  content_osd_1_textarea_3_stream=##MEDIATITLE##